apps/backend/packages/requirements.txt
# Basic dependencies for AI core functionality
numpy
openai
pdfplumber
python-docx
qdrant-client
scikit-learn
# OCR dependencies for image text extraction
# pillow
# pytesseract